Get the epidemiological calendar of a consulted year.

Function that returns the starting date of the epidemiological weeks in a year of interest.

Usage

epi_calendar(year, jan_days = 4)

Arguments

year

A numeric value for the year of interest.

jan_days

Number of January days that the first epidemiological week must contains.

Value

A character array with the starting dates of the epidemiological weeks of the given year.

Examples

epi_calendar(2016)
#>  [1] "2016-01-03" "2016-01-10" "2016-01-17" "2016-01-24" "2016-01-31"
#>  [6] "2016-02-07" "2016-02-14" "2016-02-21" "2016-02-28" "2016-03-06"
#> [11] "2016-03-13" "2016-03-20" "2016-03-27" "2016-04-03" "2016-04-10"
#> [16] "2016-04-17" "2016-04-24" "2016-05-01" "2016-05-08" "2016-05-15"
#> [21] "2016-05-22" "2016-05-29" "2016-06-05" "2016-06-12" "2016-06-19"
#> [26] "2016-06-26" "2016-07-03" "2016-07-10" "2016-07-17" "2016-07-24"
#> [31] "2016-07-31" "2016-08-07" "2016-08-14" "2016-08-21" "2016-08-28"
#> [36] "2016-09-04" "2016-09-11" "2016-09-18" "2016-09-25" "2016-10-02"
#> [41] "2016-10-09" "2016-10-16" "2016-10-23" "2016-10-30" "2016-11-06"
#> [46] "2016-11-13" "2016-11-20" "2016-11-27" "2016-12-04" "2016-12-11"
#> [51] "2016-12-18" "2016-12-25"